# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Deriv Quant Research & Execution Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /settings:
    get:
      operationId: getSettings
      tags: [settings]
      summary: Get all configurable platform settings
      description: Returns all trading parameters with their current values (or defaults if not yet set)
      responses:
        "200":
          description: Current settings
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PlatformSettings"
    post:
      operationId: updateSettings
      tags: [settings]
      summary: Update one or more platform settings
      description: Saves changed settings to the database. Changes take effect on the next scheduler cycle.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PlatformSettings"
      responses:
        "200":
          description: Settings updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ActionResponse"
...
  /settings/ai-apply-suggestion:
    post:
      operationId: applyAiSuggestion
      tags: [settings]
      summary: Apply a pending AI suggestion to settings
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: true
      responses:
        "200":
          description: Suggestion applied
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ActionResponse"
...
components:
  schemas:
...
    ActionResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
      required:
        - success
        - message
...
    PlatformSettings:
      type: object
      properties:
        max_open_trades:
          type: string
        equity_pct_per_trade:
          type: string
        tp_multiplier_strong:
          type: string
        tp_multiplier_medium:
          type: string
        tp_multiplier_weak:
          type: string
        sl_ratio:
          type: string
        trailing_stop_pct:
          type: string
        time_exit_window_hours:
          type: string
        max_daily_loss_pct:
          type: string
        max_weekly_loss_pct:
          type: string
        max_drawdown_pct:
          type: string
        kill_switch:
          type: string
        allocation_mode:
          type: string
        total_capital:
          type: string
        scan_interval_seconds:
          type: string
        paper_equity_pct_per_trade:
          type: string
        live_equity_pct_per_trade:
          type: string
        paper_max_open_trades:
          type: string
        live_max_open_trades:
          type: string
        ai_verification_enabled:
          type: string
        deriv_api_token:
          type: string
        openai_api_key:
          type: string
        deriv_api_token_set:
          type: string
        deriv_api_token_demo:
          type: string
        deriv_api_token_demo_set:
          type: string
        deriv_api_token_real:
          type: string
        deriv_api_token_real_set:
          type: string
        openai_api_key_set:
          type: string
        trading_mode:
          type: string
        enabled_symbols:
          type: string
        paper_max_daily_loss_pct:
          type: string
        live_max_daily_loss_pct:
          type: string
        paper_max_weekly_loss_pct:
          type: string
        live_max_weekly_loss_pct:
          type: string
        paper_max_drawdown_pct:
          type: string
        live_max_drawdown_pct:
          type: string
        paper_capital:
          type: string
        demo_capital:
          type: string
        real_capital:
          type: string
        demo_equity_pct_per_trade:
          type: string
        real_equity_pct_per_trade:
          type: string
        demo_max_open_trades:
          type: string
        real_max_open_trades:
          type: string
        demo_max_daily_loss_pct:
          type: string
        real_max_daily_loss_pct:
          type: string
        demo_max_weekly_loss_pct:
          type: string
        real_max_weekly_loss_pct:
          type: string
        demo_max_drawdown_pct:
          type: string
        real_max_drawdown_pct:
          type: string
        paper_mode_active:
          type: string
        demo_mode_active:
          type: string
        real_mode_active:
          type: string
...